refactor bin/www and app.js
