+++ 01-Login/routes/auth.js
@@ -1,0 +1,61 @@
+const express = require('express');
+const passport = require('passport');
+const router = express.Router();
+const Auth0Strategy = require('passport-auth0');
+
+// This will configure Passport to use Auth0
+const strategy = new Auth0Strategy(
+  {
+    domain: process.env.AUTH0_DOMAIN,
+    clientID: process.env.AUTH0_CLIENT_ID,
+    clientSecret: process.env.AUTH0_CLIENT_SECRET,
+    callbackURL:
+      process.env.AUTH0_CALLBACK_URL || 'http://localhost:3000/callback'
+  },
+  function(accessToken, refreshToken, extraParams, profile, done) {
+    // accessToken is the token to call Auth0 API (not needed in the most cases)
+    // extraParams.id_token has the JSON Web Token
+    // profile has all the information from the user
+    return done(null, profile);
+  }
+);
+
+passport.use(strategy);
+// you can use this section to keep a smaller payload
+passport.serializeUser((user, done) => done(null, user));
+passport.deserializeUser((user, done) => done(null, user));
+
+
+router.get('/login', passport.authenticate('auth0', {
+  scope: 'openid email profile'
+}), function(req, res) {
+  res.redirect("/");
+});
+
+router.get('/logout', function(req, res) {
+  req.logout();
+  res.redirect('/');
+});
+
+router.get('/callback',
+  passport.authenticate('auth0', {
+    failureRedirect: '/failure'
+  }),
+  function(req, res) {
+    const returnTo = req.session.returnTo;
+    delete req.session.returnTo;
+    res.redirect(returnTo || '/');
+  }
+);
+
+router.get('/failure', function(req, res) {
+  var error = req.flash("error");
+  var error_description = req.flash("error_description");
+  req.logout();
+  res.render('failure', {
+    error: error[0],
+    error_description: error_description[0],
+  });
+});
+
+module.exports = router;
