+++ 01-Login/routes/home.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+/* GET home page. */
+router.get('/', function(req, res, next) {
+  res.render('index');
+});
+
+module.exports = router;
