--- 01-Login/routes/index.js
@@ -1,42 +1,0 @@
-const express = require('express');
-const passport = require('passport');
-const router = express.Router();
-
-/* GET home page. */
-router.get('/', function(req, res, next) {
-  res.render('index');
-});
-
-router.get('/login', passport.authenticate('auth0', {
-  scope: 'openid email profile'
-}), function(req, res) {
-  res.redirect("/");
-});
-
-router.get('/logout', function(req, res) {
-  req.logout();
-  res.redirect('/');
-});
-
-router.get('/callback',
-  passport.authenticate('auth0', {
-    failureRedirect: '/failure'
-  }),
-  function(req, res) {
-    const returnTo = req.session.returnTo;
-    delete req.session.returnTo;
-    res.redirect(returnTo || '/');
-  }
-);
-
-router.get('/failure', function(req, res) {
-  var error = req.flash("error");
-  var error_description = req.flash("error_description");
-  req.logout();
-  res.render('failure', {
-    error: error[0],
-    error_description: error_description[0],
-  });
-});
-
-module.exports = router;
