+++ 01-Login/routes/user.js
@@ -3,7 +3,7 @@ const protected = require('../lib/middleware/protected');
 const router = express.Router();
 
 /* GET user profile. */
-router.get('/', protected(), function(req, res, next) {
+router.get('/user', protected(), function(req, res, next) {
   res.render('user', {
     user: req.user ,
     userProfile: JSON.stringify(req.user, null, '  ')
